Finish implementation of ConfigFileLoader and ConfigFileCredentialProvider
