+++ Sources/SotoCore/Credential/CredentialProvider.swift
@@ -108,9 +108,9 @@ extension CredentialProviderFactory {
     }
 
     /// Use this method to load credentials from your AWS cli credentials and optional profile configuration files, normally located at `~/.aws/credentials` and `~/.aws/config`.
-    public static func configFile(credentialsFilePath: String = "~/.aws/credentials", configFilePath: String = "~/.aws/config", profile: String? = nil) -> CredentialProviderFactory {
+    public static func configFile(credentialsFilePath: String = "~/.aws/credentials", configFilePath: String? = nil, profile: String? = nil) -> CredentialProviderFactory {
         return Self { context in
-            let provider = AWSConfigFileCredentialProvider(credentialsFilePath: credentialsFilePath, configFilePath: configFilePath, profile: profile, context: context)
+            let provider = ConfigFileCredentialProvider(credentialsFilePath: credentialsFilePath, configFilePath: configFilePath, profile: profile, context: context)
             return DeferredCredentialProvider(context: context, provider: provider)
         }
     }
