+++ Tests/SotoCoreTests/Credential/ConfigFileLoaderTests.swift
@@ -21,7 +21,121 @@ import XCTest
 
 class ConfigFileLoadersTests: XCTestCase {
 
-    // MARK: - Config File
+    // MARK: Shared Credentials parsing (combined credentials & config)
+
+    func makeContext() throws -> CredentialProviderFactory.Context {
+        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 1)
+        defer { XCTAssertNoThrow(try eventLoopGroup.syncShutdownGracefully()) }
+        let eventLoop = eventLoopGroup.next()
+        let httpClient = HTTPClient(eventLoopGroupProvider: .shared(eventLoop))
+        defer { XCTAssertNoThrow(try httpClient.syncShutdown()) }
+
+        return .init(httpClient: httpClient, eventLoop: eventLoop, logger: TestEnvironment.logger)
+    }
+
+    func testConfigFileCredentials() {
+        let profile = "profile1"
+        let accessKey = "FAKE-ACCESS-KEY123"
+        let secretKey = "Asecretreglkjrd"
+        let sessionToken = "xyz"
+        let credential = """
+        [\(profile)]
+        aws_access_key_id=\(accessKey)
+        aws_secret_access_key=\(secretKey)
+        aws_session_token=\(sessionToken)
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        var cred: CredentialProvider?
+        XCTAssertNoThrow(cred = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: profile, context: makeContext()))
+
+        XCTAssertEqual((cred as? StaticCredential)?.accessKeyId, accessKey)
+        XCTAssertEqual((cred as? StaticCredential)?.secretAccessKey, secretKey)
+        XCTAssertEqual((cred as? StaticCredential)?.sessionToken, sessionToken)
+    }
+
+    func testConfigFileCredentialsMissingAccessKey() {
+        let profile = "profile1"
+        let secretKey = "Asecretreglkjrd"
+        let credential = """
+        [\(profile)]
+        aws_secret_access_key=\(secretKey)
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        XCTAssertThrowsError(_ = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: profile, context: makeContext())) {
+            XCTAssertEqual($0 as? ConfigFileLoader.ConfigFileError, .missingAccessKeyId)
+        }
+    }
+
+    func testConfigFileCredentialsMissingSecretKey() {
+        let profile = "profile1"
+        let accessKey = "FAKE-ACCESS-KEY123"
+        let credential = """
+        [\(profile)]
+        aws_access_key_id=\(accessKey)
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        XCTAssertThrowsError(_ = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: profile, context: makeContext())) {
+            XCTAssertEqual($0 as? ConfigFileLoader.ConfigFileError, .missingSecretAccessKey)
+        }
+    }
+
+    func testConfigFileCredentialsMissingSessionToken() {
+        let profile = "profile1"
+        let accessKey = "FAKE-ACCESS-KEY123"
+        let secretKey = "Asecretreglkjrd"
+        let credential = """
+        [\(profile)]
+        aws_access_key_id=\(accessKey)
+        aws_secret_access_key=\(secretKey)
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        var cred: CredentialProvider?
+        XCTAssertNoThrow(cred = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: profile, context: makeContext()))
+
+        XCTAssertEqual((cred as? StaticCredential)?.accessKeyId, accessKey)
+        XCTAssertEqual((cred as? StaticCredential)?.secretAccessKey, secretKey)
+        XCTAssertNil((cred as? StaticCredential)?.sessionToken)
+    }
+
+    func testConfigFileCredentialsMissingProfile() {
+        let profile = "profile1"
+        let accessKey = "FAKE-ACCESS-KEY123"
+        let secretKey = "Asecretreglkjrd"
+        let credential = """
+        [\(profile)]
+        aws_access_key_id=\(accessKey)
+        aws_secret_access_key=\(secretKey)
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        XCTAssertThrowsError(_ = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: "profile2", context: makeContext())) {
+            XCTAssertEqual($0 as? ConfigFileLoader.ConfigFileError, .missingProfile("profile2"))
+        }
+    }
+
+    func testConfigFileCredentialsParseFailure() {
+        let credential = """
+        [default]
+        aws_access_key_id
+        """
+
+        var byteBuffer = ByteBufferAllocator().buffer(capacity: credential.utf8.count)
+        byteBuffer.writeString(credential)
+        XCTAssertThrowsError(_ = try ConfigFileLoader.sharedCredentials(from: byteBuffer, for: "default", context: makeContext())) {
+            XCTAssertEqual($0 as? ConfigFileLoader.ConfigFileError, .invalidCredentialFileSyntax)
+        }
+    }
+
+    // MARK: - Config File parsing
 
     func testConfigFileDefault() throws {
         let content = """
@@ -36,7 +150,7 @@ class ConfigFileLoadersTests: XCTestCase {
         let config = try ConfigFileLoader.loadProfileConfig(from: byteBuffer, for: ConfigFileLoader.default)
         XCTAssertEqual(config.roleArn, "arn:aws:iam::123456789012:role/marketingadminrole")
         XCTAssertEqual(config.sourceProfile, "user1")
-        XCTAssertEqual(config.region, "us-west-2")
+        XCTAssertEqual(config.region, .uswest2)
     }
 
     func testConfigFileNamedProfile() throws {
@@ -58,7 +172,7 @@ class ConfigFileLoadersTests: XCTestCase {
         XCTAssertEqual(config.roleArn, "arn:aws:iam::123456789012:role/marketingadminrole")
         XCTAssertEqual(config.sourceProfile, "user1")
         XCTAssertEqual(config.roleSessionName, "foo@example.com")
-        XCTAssertEqual(config.region, "us-west-1")
+        XCTAssertEqual(config.region, .uswest1)
     }
 
     func testConfigFileCredentialSourceEc2() throws {
@@ -102,7 +216,7 @@ class ConfigFileLoadersTests: XCTestCase {
         XCTAssertEqual(config.credentialSource, .environment)
     }
 
-    // MARK: - Credentials File
+    // MARK: - Credentials File parsing
 
     func testCredentialsDefault() throws {
         let content = """
@@ -157,34 +271,4 @@ class ConfigFileLoadersTests: XCTestCase {
         XCTAssertEqual(config.sourceProfile, ConfigFileLoader.default)
     }
 
-    // MARK: - Config file path expansion
-
-    func testExpandTildeInFilePath() {
-        let expandableFilePath = "~/.aws/credentials"
-        let expandedNewPath = ConfigFileLoader.expandTildeInFilePath(expandableFilePath)
-
-        #if os(Linux)
-        XCTAssert(!expandedNewPath.hasPrefix("~"))
-        #else
-
-        #if os(macOS)
-        // on macOS, we want to be sure the expansion produces the posix $HOME and
-        // not the sanboxed home $HOME/Library/Containers/<bundle-id>/Data
-        let macOSHomePrefix = "/Users/"
-        XCTAssert(expandedNewPath.starts(with: macOSHomePrefix))
-        XCTAssert(!expandedNewPath.contains("/Library/Containers/"))
-        #endif
-
-        // this doesn't work on linux because of SR-12843
-        let expandedNSString = NSString(string: expandableFilePath).expandingTildeInPath
-        XCTAssertEqual(expandedNewPath, expandedNSString)
-        #endif
-
-        let unexpandableFilePath = "/.aws/credentials"
-        let unexpandedNewPath = ConfigFileLoader.expandTildeInFilePath(unexpandableFilePath)
-        let unexpandedNSString = NSString(string: unexpandableFilePath).expandingTildeInPath
-
-        XCTAssertEqual(unexpandedNewPath, unexpandedNSString)
-        XCTAssertEqual(unexpandedNewPath, unexpandableFilePath)
-    }
 }
